Moved the code of the method "SelectTabAsync" to the page navigation service to make it mockable
